[reorg] Move LigatureSubst to new layout.
