+++ src/Makefile.sources
@@ -98,6 +98,10 @@ HB_BASE_sources = \
 	OT/Layout/GSUB/AlternateSubstFormat1.hh \
 	OT/Layout/GSUB/AlternateSubst.hh \
 	OT/Layout/GSUB/AlternateSet.hh \
+	OT/Layout/GSUB/LigatureSubstFormat1.hh \
+	OT/Layout/GSUB/LigatureSubst.hh \
+	OT/Layout/GSUB/LigatureSet.hh \
+	OT/Layout/GSUB/Ligature.hh \
 	hb-ot-layout-gsubgpos.hh \
 	hb-ot-layout-jstf-table.hh \
 	hb-ot-layout.cc \
